Jal/Jalr problems solved, Jalr target address modified
I modify csrc/Makefile to generate .elf file and use
riscv-non-elf-objdump to see the detailed instructions in fibonacci.c.
After that, I use verilator and surfer to observe signals and debug
codes.

The first issue is the stored PC calculation of Jal/Jalr. I forget to make
PC+4.
The second issue is the target PC calculation of Jalr. I mistake the 1.U
and 1.U(32.W).